phy: dp8382x: Revert the phy_write_mmd to legacy call
In the linux 4.9 the phy_write_mmd does not allow non-c45
devices to do extended register access through the phy_write_mmd
call.  This was done with the phy_write_mmd_indirect call.

The phy_write_mmd was updated in Linux 4.12 to allow either c45
or other clause device to use the standard phy_write_mmd call
and that function made the appropriate calls.

It is easier to modify the driver to conform to the current kernel
then attempt to backport numerous patches that may break other drivers.

Signed-off-by: Dan Murphy <dmurphy@ti.com>