net: phy: DP83822: Add setting the fixed internal delay
commit 8095295292b564583bb5bee5ba69924c2fd40434 from upstream

The DP83822 can be configured to use the RGMII interface. There are
independent fixed 3.5ns clock shift (aka internal delay) for the TX and RX
paths. This allow either one to be set if the MII interface is RGMII and
the value is set in the firmware node.

Signed-off-by: Dan Murphy <dmurphy@ti.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 8095295292b564583bb5bee5ba69924c2fd40434)